+++ src/component-resolver/index.js
@@ -1,0 +1,3 @@
+import resolvePath from './components-resolver';
+
+export default resolvePath;